Add status file and imported as modules
